pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diaxiinject"
version = "0.1.0"
description = "LLM security testing tool - uses local uncensored LLMs to test cloud LLM security"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.50.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "sqlite-utils>=3.36",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
]
gpu = [
    "vllm>=0.6.0",
    "torch>=2.4.0",
]

[project.scripts]
diaxiinject = "diaxiinject.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/diaxiinject"]

[tool.ruff]
target-version = "py311"
line-length = 100
